requirements.txt
streamlit>=1.32.0
supabase>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.6.0
mailjet-rest>=1.3.4
pytest>=7.4.0  # Added for running tests
PyYAML>=6.0.1  # Added for config.yaml handling
email-validator>=2.1.0  # Added for EmailStr validation
reportlab>=4.0.8  # For PDF generation
anthropic>=0.7.0  # For AI-assisted account classification
